+++ sdk/src/main/java/com/eyeson/sdk/model/local/meeting/CustomMessage.kt
@@ -1,0 +1,11 @@
+package com.eyeson.sdk.model.local.meeting
+
+import com.eyeson.sdk.model.local.base.LocalBaseCommand
+import java.util.*
+
+internal data class CustomMessage(
+    val content: String,
+    val createdAt: Date,
+    val userId: String
+) : LocalBaseCommand
+