+++ sdk/src/main/java/com/eyeson/sdk/model/sepp/outgoing/SeppCommandsOutgoing.kt
@@ -4,6 +4,10 @@ package com.eyeson.sdk.model.sepp.outgoing
 internal enum class SeppCommandsOutgoing(val type: String) {
     CALL_START("call_start"),
     CALL_TERMINATE("call_terminate"),
-    CALL_RESUME("call_resume")
+    CALL_RESUME("call_resume"),
+    CHAT_OUTGOING("chat"),
+    SET_PRESENTER("set_presenter"),
+    DESKTOP_STREAMING("desktopstreaming"),
+    MUTE_VIDEO("mute_video")
 }