+++ sdk/src/main/java/com/eyeson/sdk/moshi/adapter/SeppIncomingCommandsAdapter.kt
@@ -6,11 +6,11 @@ import com.eyeson.sdk.model.sepp.incoming.CallAcceptedDto
 import com.eyeson.sdk.model.sepp.incoming.CallRejectedDto
 import com.eyeson.sdk.model.sepp.incoming.CallResumedDto
 import com.eyeson.sdk.model.sepp.incoming.CallTerminatedDto
+import com.eyeson.sdk.model.sepp.incoming.ChatIncomingDto
+import com.eyeson.sdk.model.sepp.incoming.MemberListUpdateDto
+import com.eyeson.sdk.model.sepp.incoming.RecordingStatusDto
 import com.eyeson.sdk.model.sepp.incoming.SeppCommandsIncoming
-import com.eyeson.sdk.model.sepp.outgoing.CallResumeDto
-import com.eyeson.sdk.model.sepp.outgoing.CallStartDto
-import com.eyeson.sdk.model.sepp.outgoing.CallTerminateDto
-import com.eyeson.sdk.model.sepp.outgoing.SeppCommandsOutgoing
+import com.eyeson.sdk.model.sepp.incoming.SourceUpdateDto
 import com.eyeson.sdk.model.sepp.shared.SdpUpdateDto
 import com.eyeson.sdk.model.sepp.shared.SeppCommandsShared
 import com.squareup.moshi.adapters.PolymorphicJsonAdapterFactory
@@ -23,10 +23,17 @@ internal object SeppIncomingCommandsAdapter {
             .withSubtype(CallRejectedDto::class.java, SeppCommandsIncoming.CALL_REJECTED.type)
             .withSubtype(CallTerminatedDto::class.java, SeppCommandsIncoming.CALL_TERMINATED.type)
             .withSubtype(CallResumedDto::class.java, SeppCommandsIncoming.CALL_RESUMED.type)
-            .withSubtype(CallStartDto::class.java, SeppCommandsOutgoing.CALL_START.type)
-            .withSubtype(CallTerminateDto::class.java, SeppCommandsOutgoing.CALL_TERMINATE.type)
-            .withSubtype(CallResumeDto::class.java, SeppCommandsOutgoing.CALL_RESUME.type)
             .withSubtype(SdpUpdateDto::class.java, SeppCommandsShared.SDP_UPDATE.type)
+            .withSubtype(ChatIncomingDto::class.java, SeppCommandsIncoming.CHAT.type)
+            .withSubtype(SourceUpdateDto::class.java, SeppCommandsIncoming.SOURCE_UPDATE.type)
+            .withSubtype(
+                MemberListUpdateDto::class.java,
+                SeppCommandsIncoming.MEMBERLIST_UPDATE.type
+            )
+            .withSubtype(
+                RecordingStatusDto::class.java,
+                SeppCommandsIncoming.RECORDING_STATUS.type
+            )
             .withDefaultValue(UnknownCommandDto())
     }
 }