+++ sdk/src/main/java/com/eyeson/sdk/network/WebSocketCommunicator.kt
@@ -11,6 +11,10 @@ import com.eyeson.sdk.model.local.sepp.CallAccepted
 import com.eyeson.sdk.model.local.sepp.CallResume
 import com.eyeson.sdk.model.local.sepp.CallStart
 import com.eyeson.sdk.model.local.sepp.CallTerminate
+import com.eyeson.sdk.model.local.sepp.ChatOutgoing
+import com.eyeson.sdk.model.local.sepp.DesktopStreaming
+import com.eyeson.sdk.model.local.sepp.MuteVideo
+import com.eyeson.sdk.model.local.sepp.SetPresenter
 import com.eyeson.sdk.model.local.ws.ReconnectSignaling
 import com.eyeson.sdk.model.local.ws.WsFailure
 import com.eyeson.sdk.model.local.ws.WsOpen
@@ -18,6 +22,10 @@ import com.eyeson.sdk.model.meeting.outgoing.MuteAllDto
 import com.eyeson.sdk.model.sepp.outgoing.CallResumeDto
 import com.eyeson.sdk.model.sepp.outgoing.CallStartDto
 import com.eyeson.sdk.model.sepp.outgoing.CallTerminateDto
+import com.eyeson.sdk.model.sepp.outgoing.ChatOutgoingDto
+import com.eyeson.sdk.model.sepp.outgoing.DesktopStreamingDto
+import com.eyeson.sdk.model.sepp.outgoing.MuteVideoDto
+import com.eyeson.sdk.model.sepp.outgoing.SetPresenterDto
 import com.eyeson.sdk.model.sepp.outgoing.fromLocal
 import com.eyeson.sdk.utils.collectIn
 import kotlinx.coroutines.CoroutineScope
@@ -84,6 +92,37 @@ internal class WebSocketCommunicator(
         meetingCommunicator?.disconnect()
         signalingConnection?.disconnect()
     }
+
+    fun sendChatMessage(message: String) {
+        val chatOutgoing = ChatOutgoing(message, meeting.signaling.options.clientId)
+        val adapter = moshi.adapter(ChatOutgoingDto::class.java)
+
+        signalingConnection?.sendMessage(adapter.toJson(chatOutgoing.fromLocal(callId, meeting)))
+    }
+
+    fun enablePresentation(enable: Boolean) {
+        val setPresenter = SetPresenter(on = enable, cid = meeting.signaling.options.clientId)
+        val desktopStreaming =
+            DesktopStreaming(on = enable, cid = meeting.signaling.options.clientId)
+
+        signalingConnection?.sendMessage(
+            moshi.adapter(SetPresenterDto::class.java)
+                .toJson(setPresenter.fromLocal(callId, meeting))
+        )
+        signalingConnection?.sendMessage(
+            moshi.adapter(DesktopStreamingDto::class.java)
+                .toJson(desktopStreaming.fromLocal(callId, meeting))
+        )
+    }
+
+
+    fun setLocalVideoEnabled(enable: Boolean) {
+        val muteVideo = MuteVideo(muted = !enable, cid = meeting.signaling.options.clientId)
+        val adapter = moshi.adapter(MuteVideoDto::class.java)
+
+        signalingConnection?.sendMessage(adapter.toJson(muteVideo.fromLocal(callId, meeting)))
+    }
+
 
     private suspend fun connectToMeetingWs(): MeetingConnection {
         return coroutineScope {